requirements.txt
# Core dependencies
Pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.0.0
openpyxl>=3.1.0

# File monitoring
watchdog>=3.0.0

# Data validation and configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0

# CLI and utilities
click>=8.1.0
rich>=13.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.5.0
black>=23.7.0
isort>=5.12.0

# Logging
structlog>=23.1.0